Worked on DirectoryNav.py
1. worked on selectCSV
2. decided on goals and built a todo list
3. tested quite a few methods and solidified some concepts in how python is set up.

Little Journal Entry On My Python Journey-
No hate, but sometimes it feels like this program is like writing a book using one or two syllable words. The for loops are really making my head spin when they definitely should not be. The way it automates datatypes confounds me and its taking some time to get used to. Overall, I'm liking the language and how easy it is to manipulate Databases with SQLite. Excited to work more with some more complex algorithms and to hopefully use some OOP languages to really open up the throttle on some of this stuff!